Add dismissal as a prop to EventForm
