+++ components/eventForm/SubmitButton.tsx
@@ -1,43 +1,6 @@
 import React from "react"
 import { Alert, Text, TouchableOpacity } from "react-native"
-import { MaterialIcons } from "@expo/vector-icons"
 import { eventEditInputFromFormValues, useEventFormContext } from "."
-
-/**
- * Props for `EventFormDismissButton`.
- */
-export type EventFormDismissButtonProps = {
-  onDismiss: () => void
-}
-
-/**
- * A dismiss button for `EventForm`.
- *
- * If the user has edited any of the form values, this button will display
- * a confirmation alert before dismissing.
- */
-export const EventFormDismissButton = ({
-  onDismiss
-}: EventFormDismissButtonProps) => {
-  const { hasEdited } = useEventFormContext()
-
-  const dismissButtonTapped = () => {
-    if (hasEdited) {
-      Alert.alert("Discard this draft?", undefined, [
-        { text: "Discard", style: "cancel", onPress: onDismiss },
-        { text: "Keep Editing" }
-      ])
-    } else {
-      onDismiss()
-    }
-  }
-
-  return (
-    <TouchableOpacity onPress={dismissButtonTapped} accessibilityLabel="Cancel">
-      <MaterialIcons name="close" size={24} />
-    </TouchableOpacity>
-  )
-}
 
 /**
  * Props from `EventFormSubmitButton`.
